This is more or less a rewrite of get_strings.php. It is a combination of the functionality in rev. 1.25 and the get_strings.php that i have earlier provided as a patch (unfortunately both contained errors).
The features are documeted in the file. Below comes a list of bus that have been
corrected in both the pathc and the CVS version.

Unfortunately there is no list of introduced errors but i hope it is small...


- extraction of strings is now done inside both singel and double quotes
(applies to my version of getstrings.php).

- extracted strings from smarty-files and php files is separated - since we no try
to filter out comments so that the extraction is not confused. This is done
completely in .tpl files but unfortuantely only partially in .php files
(calling php -w woudl fix this - but safe mode prevets it). Also the regular
expressions that extract the language strings now only works on the corresponding
filetype.

- All strings are converted (escaped) to the format that they would get if they
where read by 'require("langauga.php");', since then the strings are escaped
again when a new langauge.php is created. Not that this treatment is different for
 all three types of extracted strings ({tr}...{/tr}, (haw)tra ('...'); and
(haw)tra ("...");)

- In .php-file only strings inside functions tra and hawtra are now extracted
(earlier any function ending in tra with only one string parameter got strings
extracted).

- The regular expressions that extracts strings are more 'forgiving' by accepting
whitespace in the same places as PHP does.

Improvement.
- The close function is now more agressive by lowercasing all text before calculating
the simmilarity (and also strings larger than 255 characters are also treated since
they are truncated to the lenght that the lehvenstein function accepts).