[MOD] dates and timezones : Modify every call to php native date / timezone functions in order to use GMT timezone instead of the server timezone.
Timestamps stored in the database should be stored in GMT timezone, because changing the server (or just its timezone settings) should never have an impact on tikiwiki. This is why every date(), strftime() and mktime() calls are replaced by gmdate(), gmstrftime() and gmmktime(). Those functions will use GMT timestamps and not a timestamp that depends on the server TZ environment variable.

!! Note that everything has not been fully tested yet and that this update lacks of an upgrade path for existing calendars and timezones stored in database with a server timezone.