[FIX] dates and timezones : apply the user timezone on timestamps retrieved from database in tikilib.
Since all timestamps are now in GMT timezone in database, we need to convert those timestamps using the user timezone (probably in every case) before using it (displaying, comparing, ranking, ...). This is something that is not done in smarty functions anymore because it should be done only once, just after retrieving the timestamp from database. Note that this has to be used only when a select query is done (nothing should be modified when updating or inserting values in database).

This fix concerned tables are 'tiki_pages' and 'tiki_page_drafts'. There is probably more fix to commit for other tables that includes timestamps.

In order to easily convert every timestamp fields in each function that retrieve this kind of data in tikilib, a new function 'page_info_UTC_to_display()' has been added and will take care of the timestamp field list.