[QT] Quality Team Backport
r30275 | sept_7 | 2010-10-25 15:09:47 +0200 (Mon, 25 Oct 2010) | 33 lines
[FIX/bp 30171] WebDAV : fix an authentication pb with apache when PHP is used
in FCGI or CGI mode.
Extra documentation :
If Apache is calling PHP in CGI mode, authentication HTTP Headers are not
set.
In this case, you have to add the following lines inside your Apache
VirtualHost config :
   RewriteEngine on
   RewriteCond %{HTTP:Authorization} ^(.*)
   RewriteRule ^(.*) - [E=HTTP_AUTHORIZATION:%1]
Thanx to Changi that points me to a fix by Nyloth (I suppose ;) )
Now this rules are in _htaccess
[FIX/bp 30201] WebDAV and HTTP(S) authentication with Apache in CGI or FCGI
mode
[FIX/ bp 30212] WebDAV : Workaroung tiki-webdav always
assumes we are always in HTTP mode and ignore HTTPS mode :( 
[FIX/bp 30264] Use SCRIPT_NAME instead of PHP_SELF : homogenize use of
SCRIPT_NAME in tiki-filter-base.php as it is used all over this script and
PHP_SELF is used only twice... 
[FIX/bp 30265] WebDAV : Simplify path handling in WebDAV, revert 30212 as the
handling was suspicious from the start, it was not an ezcomponent bug, 
IMPORTANT NOTE : needs 30264 to work 
[ENH/bp 30266] WebDAV : better message for webdav link in file gallery
All [FIX] patches are required to fix the different pb with WebDAV, 
that's why I submit them in batch...